int -> size_t 로 변경

+++ ImageBuffer_Test/ImageFileUtil.h
@@ -5,7 +5,7 @@
 
 using byte = unsigned char;
 
-void ReadImageFileHeader(std::wstring imageFilePath, int* pWidth, int* pHeight, int* pStride) {
+void ReadImageFileHeader(std::wstring imageFilePath, size_t* pWidth, size_t* pHeight, size_t* pStride) {
     // todo
     *pWidth = 64;
     *pHeight = 64;
